# Changes (Ptr Word8, a) to PeekResult {-# UNPACK #-} (Ptr Word8) !a